requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
pymongo==4.6.3
mongoengine==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
python-dotenv==1.0.0
pydantic~=2.10.4